requirements.txt
Flask
requests
PyYAML
